+++ modules/library/main/src/test/java/org/geotools/filter/IsEqualsToImplTest.java
@@ -101,8 +101,7 @@ public class IsEqualsToImplTest {
 
     @Test
     public void testOperandsDoubleLongOutOfRange() {
-        Expression e1 =
-                filterFactory.literal(Double.valueOf(Long.MAX_VALUE).doubleValue() + 10000.0);
+        Expression e1 = filterFactory.literal(Long.MAX_VALUE + 10000d);
         Expression e2 = filterFactory.literal(Long.MAX_VALUE);
 
         PropertyIsEqualTo equal = filterFactory.equals(e1, e2);
